rename more tests
